refactor(voice-chat): move voiceChatMachine + its tests into the service
Co-locates the xstate machine with the wrapping service per spec File-
structure decision. The pure xstate value semantics are unchanged; the 12
existing machine tests are kept verbatim with one relative-import swap.

Deviation: a re-export shim is kept at machines/voiceChatMachine.ts until
Task 15+ migrates the legacy voiceChatService.ts and chatStore.ts off the
old import path. The shim will be deleted in Task 16.
